// client/src/app/openapi/model/backupDirectory.ts
/**
 * BackupHelper.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {BackupFile} from './backupFile';


export interface BackupDirectory {
    name?: string | null;
    directories?: Array<BackupDirectory> | null;
    files?: Array<BackupFile> | null;
}